locall working rag test this/requirements.txt
fastapi[all]
uvicorn
chromadb
openai
requests
python-dotenv
aiohttp
langchain-text-splitters
